+++ R/modelStudio.R
@@ -186,7 +186,7 @@ modelStudio.default <- function(object,
   fi <- try_catch(
     ingredients::feature_importance(
         model, data, y, predict_function, variables = variable_names, B = B),
-    "ingredients::feature_importance")
+    "ingredients::feature_importance", "", show_info)
 
   if (show_info) setTxtProgressBar(pb, 1)
 
@@ -197,13 +197,13 @@ modelStudio.default <- function(object,
     pd_n <- try_catch(
       ingredients::partial_dependency(
           model, data, predict_function, variable_type = "numerical", N = N),
-      "ingredients::partial_dependency")
+      "ingredients::partial_dependency", "numerical", show_info)
     if (show_info) setTxtProgressBar(pb, 2)
     pd_c <- NULL
     ad_n <- try_catch(
       ingredients::accumulated_dependency(
           model, data, predict_function, variable_type = "numerical", N = N),
-      "ingredients::accumulated_dependency")
+      "ingredients::accumulated_dependency", "numerical", show_info)
     if (show_info) setTxtProgressBar(pb, 4)
     ad_c <- NULL
   } else if (all(!which_numerical)) {
@@ -211,34 +211,34 @@ modelStudio.default <- function(object,
     pd_c <- try_catch(
       ingredients::partial_dependency(
           model, data, predict_function, variable_type = "categorical", N = N),
-      "ingredients::partial_dependenct")
+      "ingredients::partial_dependenct", "categorical", show_info)
     if (show_info) setTxtProgressBar(pb, 3)
     ad_n <- NULL
     ad_c <- try_catch(
       ingredients::accumulated_dependency(
           model, data, predict_function, variable_type = "categorical", N = N),
-      "ingredients::accumulated_dependency")
+      "ingredients::accumulated_dependency", "categorical", show_info)
     if (show_info) setTxtProgressBar(pb, 5)
   } else {
     pd_n <- try_catch(
       ingredients::partial_dependency(
         model, data, predict_function, variable_type = "numerical", N = N),
-      "ingredients::partial_dependency")
+      "ingredients::partial_dependency", "numerical", show_info)
     if (show_info) setTxtProgressBar(pb, 2)
     pd_c <- try_catch(
       ingredients::partial_dependency(
         model, data, predict_function, variable_type = "categorical", N = N),
-      "ingredients::partial_dependenct")
+      "ingredients::partial_dependenct", "categorical", show_info)
     if (show_info) setTxtProgressBar(pb, 3)
     ad_n <- try_catch(
       ingredients::accumulated_dependency(
         model, data, predict_function, variable_type = "numerical", N = N),
-      "ingredients::accumulated_dependency")
+      "ingredients::accumulated_dependency", "numerical", show_info)
     if (show_info) setTxtProgressBar(pb, 4)
     ad_c <- try_catch(
       ingredients::accumulated_dependency(
         model, data, predict_function, variable_type = "categorical", N = N),
-      "ingredients::accumulated_dependency")
+      "ingredients::accumulated_dependency", "categorical", show_info)
     if (show_info) setTxtProgressBar(pb, 5)
   }
 
@@ -247,6 +247,7 @@ modelStudio.default <- function(object,
   ad_data <- prepare_accumulated_dependency(ad_n, ad_c, variables = variable_names)
   fd_data <- prepare_feature_distribution(data, variables = variable_names)
   tv_data <- prepare_target_vs(data, y, variables = variable_names)
+  ta_data <- prepare_target_average(data, y, variables = variable_names)
 
   if (parallel) {
     parallelMap::parallelStart()
@@ -258,15 +259,15 @@ modelStudio.default <- function(object,
       bd <- try_catch(
         iBreakDown::local_attributions(
           model, data, predict_function, new_observation, label = label),
-        "iBreakDown::local_attributions")
+        "iBreakDown::local_attributions", i, show_info)
       sv <- try_catch(
         iBreakDown::shap(
           model, data, predict_function, new_observation, label = label, B = B),
-        "iBreakDown::shap")
+        "iBreakDown::shap", i, show_info, FALSE)
       cp <- try_catch(
         ingredients::ceteris_paribus(
           model, data, predict_function, new_observation, label = label),
-        "ingredients::ceteris_paribus")
+        "ingredients::ceteris_paribus", i, show_info, FALSE)
 
       bd_data <- prepare_break_down(bd, max_features, ...)
       sv_data <- prepare_shap_values(sv, max_features, ...)
@@ -296,15 +297,15 @@ modelStudio.default <- function(object,
       bd <- try_catch(
         iBreakDown::local_attributions(
           model, data, predict_function, new_observation, label = label),
-        "iBreakDown::local_attributions")
+        "iBreakDown::local_attributions", i, show_info)
       sv <- try_catch(
         iBreakDown::shap(
           model, data, predict_function, new_observation, label = label, B = B),
-        "iBreakDown::shap")
+        "iBreakDown::shap", i, show_info, FALSE)
       cp <- try_catch(
         ingredients::ceteris_paribus(
           model, data, predict_function, new_observation, label = label),
-        "ingredients::ceteris_paribus")
+        "ingredients::ceteris_paribus", i, show_info, FALSE)
 
       if (show_info) setTxtProgressBar(pb, 5 + i)
 
@@ -335,7 +336,7 @@ modelStudio.default <- function(object,
                     drop_down_data = jsonlite::toJSON(drop_down_data)
                     ), options)
 
-  temp <- jsonlite::toJSON(list(obs_list, fi_data, pd_data, ad_data, fd_data, tv_data))
+  temp <- jsonlite::toJSON(list(obs_list, fi_data, pd_data, ad_data, fd_data, tv_data, ta_data))
 
   sizing_policy <- r2d3::sizingPolicy(padding = 10, browser.fill = TRUE)
 
@@ -403,13 +404,18 @@ remove_file_paths <- function(text, type = NULL) {
 #'
 #' @param expr function
 #' @param function_name string
+#' @param i iteration/type
+#' @param show_info show message about what is calculated
 #'
 #' @return Valid object or \code{NULL}
 
-try_catch <- function(expr, function_name) {
-  tryCatch(expr,
-           error = function(e) {
-             warning(paste0("Error occurred in ", function_name, " function: ", e$message))
-             NULL
-           })
+try_catch <- function(expr, function_name, i = 1, show_info = TRUE, new = TRUE) {
+  tryCatch({
+    if(show_info) message(paste0(ifelse(new,"\n",""), "Function ", function_name, " is beeing calculated. (", i, ")"))
+    expr
+    },
+    error = function(e) {
+      warning(paste0("Error occurred in ", function_name, " function: ", e$message))
+      NULL
+  })
 }